--- backend/vercel.json
@@ -1,16 +1,0 @@
-{
-    "version": 2,
-    "builds": [
-      {
-        "src": "server.ts",
-        "use": "@vercel/node"
-      }
-    ],
-    "routes": [
-      {
-        "src": "/(.*)",
-        "dest": "/server.ts"
-      }
-    ]
-  }
-
