+++ frontend/src/components/BackgroundContext/BackgroundContext.css
@@ -8,7 +8,7 @@
   }
   
   .flash-green {
-    background-color: rgba(41, 219, 41, 0.571);
+    background-color: rgba(41, 219, 41, 0.74);
   }
   
   .flash-red {
