+++ frontend/src/components/appRoutes/index.tsx
@@ -4,7 +4,9 @@ import SoloGame from '../soloGame';
 import CreateUser from '../createUserPage';
 import LoginPage from '../loginPage';
 import NotFound from '../notFound';
+import RoomForm from '../roomForm';
 import CoopRoom from '../coopRoom';
+
 
 const AppRoutes: React.FC = () => {
 	return (
@@ -14,7 +16,7 @@ const AppRoutes: React.FC = () => {
 			<Route path="/register" element={<CreateUser/>} />
 			<Route path="/login" element={<LoginPage />} />
 			<Route path="/soloGame" element={<SoloGame />} />
-			<Route path="/coopRoom" element={<CoopRoom />} />
+			<Route path="/coop" element={<CoopRoom />} />
 		</Routes>
 	);
 };
