+++ frontend/src/components/coopRoom/index.tsx
@@ -1,75 +1,28 @@
-import { useState } from 'react'
-import { Link } from 'react-router-dom'
-import './coopRoom.css'
-import axios from 'axios'
-import SnackBar from '../snackbar'
+
+import { useState } from "react";
+import RoomForm from "../roomForm";
+import ButtonLink from "../buttonLink";
+
 
 const CoopRoom = () => {
-	const [roomName, setRoomName] = useState('')
-	const [roomCode, setRoomCode] = useState('')
-	const [registerError, setRegisterError] = useState(false)
+  const [isAuth] = useState<string | null>(localStorage.getItem("token") || null);
+  const [userId] = useState<string | null>(localStorage.getItem("userId"));
 
-	const handleSubmit = async (e: React.FormEvent) => {
-		e.preventDefault()
+  return (
+    <div>
+      <h1>Cooperativo</h1>
+      {isAuth && userId? (
+          <RoomForm />
+      ) : (
+        <div>
+          <p>Faça login primeiro</p>
+          <ButtonLink buttontext="Voltar" to="/"></ButtonLink>
+        </div>
+        
+      )}
+      
+    </div>
+  );
+};
 
-		try {
-			//   await axios.post('http://localhost:3000/user/register', {
-			//     roomName,
-			//     roomCode,
-			//   });
-			//   setRegisterError(false)
-		} catch (error: any) {
-			setRegisterError(true)
-		}
-	}
-
-	// trycatch for join room here:
-	//   try {
-
-	//   } catch (error) {
-
-	//   }
-	return (
-		<div className="create-room-container">
-			<div className="create-room-card">
-				<h1 className="create-room-title">Modo Coop</h1>
-
-				<form className="create-room-form" onSubmit={handleSubmit}>
-					<div className="form-group">
-						<label>Nome da sala</label>
-						<input
-							type="text"
-							value={roomName}
-							onChange={(e) => setRoomName(e.target.value)}
-							placeholder="Digite o nome da sala"
-						/>
-					</div>
-					<button type="submit" className="create-room-button">
-						Criar Sala
-					</button>
-				</form>
-
-				<form className="create-room-form" onSubmit={handleSubmit}>
-					<div className="form-group">
-						<label>Entrar na sala</label>
-						<input
-							type="text"
-							value={roomCode}
-							onChange={(e) => setRoomCode(e.target.value)}
-							placeholder="Digite o nome da sala:"
-						/>
-					</div>
-					<button type="submit" className="create-room-button">
-						Entrar em uma Sala
-					</button>
-				</form>
-
-				<p className="create-room-link">
-					Deseja sair? <Link to="/">Menu</Link>
-				</p>
-			</div>
-		</div>
-	)
-}
-
-export default CoopRoom
+export default CoopRoom;
