+++ frontend/src/components/homePage/index.tsx
@@ -36,7 +36,7 @@ const Home = () => {
 			<div className="GameModes">
 				<ButtonLink buttontext="Solo Game" to="/soloGame" id="Solo-Game" />
 				<ButtonLink buttontext="1 vs 1" to="/" id="One-vs-One" />
-				<ButtonLink buttontext="Co-op" to="/coopRoom" id="Co-op" />
+				<ButtonLink buttontext="Co-op" to="/coop" id="Co-op" />
 				<ButtonLink buttontext="Conta" to="/register" id="Conta" />
 			</div>
 		</div>
