+++ frontend/src/components/soloGame/SoloGame.css
@@ -1,0 +1,3 @@
+.soloGame h1{
+    color: white;
+}
